requirements.txt
infobip-api-python-client
python-escpos
pyngrok
flask
